GeminiLegion-kapa/gemini-legion-backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai>=0.7.0
python-dotenv
httpx
pydantic>=2.0
